Move files to src directory.
