+++ package.json
@@ -2,7 +2,7 @@
   "name": "farmos",
   "version": "0.1.6",
   "description": "A library for retrieving data from a farmOS server",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "test": "mocha --watch"
   },
