+++ src/parseFilter.js
